+++ NEWS.md
@@ -1,3 +1,8 @@
+# myClim 1.1.4
+* Fixed the issue with the joined series of loggers in the `mc_plot_line` function.
+* Added a new parameter `color_by_logger` to the `mc_plot_line` function.
+* Added a new parameter `logger_types` to the `mc_filter` function.
+
 # myClim 1.1.3
 * Fixed reading of TOMST Thermologger data CSV. The data format has changed in the new version of Lolly.